Divide components to Subcomponents and Schedule Routes with Time on Frontend
