+++ client-frontend/public/index.html
@@ -25,7 +25,7 @@
 
     <script src="https://cdn.socket.io/socket.io-1.0.0.js"></script>
     <script>
-      var socket = io("http://localhost:8080", { transports: ["polling"] });
+      var socket = io("http://192.168.7.33:8080", { transports: ["polling"] });
       socket.on("connect", () => {
         console.log("Connection established");
       });
